refactor(model): extract row collection helpers
