+++ ci/proof.toml
@@ -636,6 +636,7 @@ paths = [
   "crates/tokmd-model/src/aggregate.rs",
   "crates/tokmd-model/src/lib.rs",
   "crates/tokmd-model/src/module_key/**",
+  "crates/tokmd-model/src/rows.rs",
   "crates/tokmd-model/src/sorting.rs",
   "crates/tokmd-model/tests/**",
   "crates/tokmd-scan/src/lib.rs",
